Twist: dia inicial desde history, llave TWIST2_<external_id> y secretos fuera de archivos versionados
- dias_iniciales_mes de los 3 productos (Phone/Twist1/Twist2) sale de
  contract_advisors_history.dias_atraso_inicial; Twist1/Twist2 ahora persisten
  en history con su producto.
- Twist2: llave y contrato_x usan el external_id numerico; formato "TWIST2_<id>".
- Exclusiones durables en reportes/asignacion: contratos con pagare (status 1,2,3),
  mora <31, y Serlefin nunca en franja 31-60.
- Seguridad/buenas practicas: los secretos salen de docker-compose.yml (ahora
  via ${VAR} interpolado desde .env no versionado) y de .env.example (placeholders
  vacios); se anaden REPORTS_EXT_*_PASSWORD al compose. jira.md redacta el HMAC.

+++ app/database/models.py
@@ -71,7 +71,7 @@ class ContractAdvisorHistory(Base):
     
     id = Column(Integer, primary_key=True, autoincrement=True)
     user_id = Column(Integer, nullable=False, index=True)
-    contract_id = Column(Integer, nullable=False, index=True)
+    contract_id = Column(BigInteger, nullable=False, index=True)
     fecha_inicial = Column("Fecha Inicial", DateTime, nullable=False)
     fecha_terminal = Column("Fecha Terminal", DateTime, nullable=True)
     tipo = Column(String(50), nullable=True, index=True)
